fix(fuse): address PR review — ABI type, read-after-truncate, rename paths, PID safety
- ffi.ts: fuse_reply_write param is size_t, use usize not u32
- mod.ts: read after O_TRUNC now checks handle.dirty to avoid returning
  stale tree data when buffer is empty
- tree.ts: rename recursively re-tracks descendant paths via retrackSubtree
  so getPath returns correct paths after directory moves
- fuse.ts: unmount verifies PID belongs to a deno/fuse process before
  sending SIGTERM, preventing accidental kills on PID reuse

+++ packages/fuse/ffi.ts
@@ -71,7 +71,7 @@ const symbols = {
     result: "i32",
   },
   fuse_reply_write: {
-    parameters: ["pointer", "u32"],
+    parameters: ["pointer", "usize"],
     result: "i32",
   },
   fuse_reply_create: {
